+++ src/main/java/com/phcworld/userservice/controller/UserCommandApiController.java
@@ -1,42 +1,27 @@
 package com.phcworld.userservice.controller;
 
-import com.phcworld.userservice.controller.port.UserResponse;
 import com.phcworld.userservice.controller.port.UserService;
+import com.phcworld.userservice.controller.response.UserResponse;
 import com.phcworld.userservice.domain.User;
-import com.phcworld.userservice.domain.port.UserRequest;
-import io.micrometer.core.annotation.Timed;
+import com.phcworld.userservice.domain.UserRequest;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.env.Environment;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 
 @RestController
 @RequestMapping("/users")
 @RequiredArgsConstructor
 @Slf4j
 @Builder
-public class UserApiController {
+public class UserCommandApiController {
 
     private final UserService userService;
-    private final Environment env;
-
-    @GetMapping("/health_check")
-    @Timed(value = "users.status", longTask = true)
-    public String status(){
-        return String.format("It's Working in User Service on PORT %s",
-                env.getProperty("local.server.port"));
-    }
 
     @ApiResponses(value = {
             @ApiResponse(responseCode = "409", description = "가입된 이메일")
@@ -47,17 +32,6 @@ public class UserApiController {
         User user = userService.register(request);
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(UserResponse.of(user));
-    }
-
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "404", description = "요청한 회원 정보 없음")
-    })
-    @GetMapping("/{userId}")
-    public ResponseEntity<UserResponse> getUserInfo(@PathVariable(name = "userId") String userId){
-        User user = userService.getUserByUserId(userId);
-        return ResponseEntity
-                .ok()
                 .body(UserResponse.of(user));
     }
 
@@ -84,28 +58,5 @@ public class UserApiController {
         return ResponseEntity
                 .ok()
                 .body(UserResponse.of(user));
-    }
-
-    @GetMapping("")
-    public ResponseEntity<Map<String, UserResponse>> getUsers(@RequestParam(value = "userIds") List<String> userIds){
-        Map<String, User> map = userService.getUsers(userIds);
-        Map<String, UserResponse> users = new HashMap<>();
-        for (String userId : map.keySet()){
-            users.put(userId, UserResponse.of(map.get(userId)));
-        }
-        return ResponseEntity
-                .ok()
-                .body(users);
-    }
-
-    @GetMapping("/search")
-    public ResponseEntity<List<UserResponse>> getUsersByName(@RequestParam(name = "name") String name){
-        List<User> users = userService.getUserByName(name);
-        List<UserResponse> responses = users.stream()
-                .map(UserResponse::of)
-                .toList();
-        return ResponseEntity
-                .ok()
-                .body(responses);
     }
 }
