+++ src/main/java/com/phcworld/userservice/controller/port/LoginService.java
@@ -1,7 +1,7 @@
 package com.phcworld.userservice.controller.port;
 
 import com.phcworld.userservice.domain.User;
-import com.phcworld.userservice.domain.port.LoginRequest;
+import com.phcworld.userservice.domain.LoginRequest;
 import com.phcworld.userservice.jwt.dto.TokenDto;
 
 public interface LoginService {
