+++ src/main/java/com/phcworld/userservice/domain/UserRequest.java
@@ -1,4 +1,4 @@
-package com.phcworld.userservice.domain.port;
+package com.phcworld.userservice.domain;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
