+++ src/main/java/com/phcworld/userservice/infrastructure/SystemUuidHolder.java
@@ -1,11 +1,12 @@
-package com.phcworld.userservice.service.port;
+package com.phcworld.userservice.infrastructure;
 
+import com.phcworld.userservice.service.port.UuidHolder;
 import org.springframework.stereotype.Component;
 
 import java.util.UUID;
 
 @Component
-public class SystemUuidHolder implements UuidHolder{
+public class SystemUuidHolder implements UuidHolder {
 
     @Override
     public String random() {
