+++ src/main/java/com/phcworld/userservice/service/LoginServiceImpl.java
@@ -2,7 +2,7 @@ package com.phcworld.userservice.service;
 
 import com.phcworld.userservice.controller.port.LoginService;
 import com.phcworld.userservice.domain.User;
-import com.phcworld.userservice.domain.port.LoginRequest;
+import com.phcworld.userservice.domain.LoginRequest;
 import com.phcworld.userservice.exception.model.NotFoundException;
 import com.phcworld.userservice.jwt.dto.TokenDto;
 import com.phcworld.userservice.security.utils.SecurityUtil;
