+++ src/test/java/com/phcworld/userservice/service/LoginServiceImplTest.java
@@ -3,7 +3,7 @@ package com.phcworld.userservice.service;
 import com.phcworld.userservice.controller.port.LoginService;
 import com.phcworld.userservice.domain.Authority;
 import com.phcworld.userservice.domain.User;
-import com.phcworld.userservice.domain.port.LoginRequest;
+import com.phcworld.userservice.domain.LoginRequest;
 import com.phcworld.userservice.exception.model.DeletedEntityException;
 import com.phcworld.userservice.exception.model.NotFoundException;
 import com.phcworld.userservice.exception.model.UnauthorizedException;
